Support ANN field also
