+++ plugins/variants/src/VcfTabixAdapter/index.ts
@@ -1,8 +1,8 @@
-export { default as configSchema } from './configSchema'
-
 import PluginManager from '@jbrowse/core/PluginManager'
 import AdapterType from '@jbrowse/core/pluggableElementTypes/AdapterType'
 import configSchema from './configSchema'
+
+export { default as configSchema } from './configSchema'
 
 export default (pluginManager: PluginManager) => {
   pluginManager.addAdapterType(
