+++ test_data/volvox/config.json
@@ -1793,11 +1793,25 @@
     {
       "type": "VariantTrack",
       "trackId": "variant_effect_demo_data",
-      "name": "Variant effect predictor example",
+      "name": "Ensembl VEP example",
       "adapter": {
         "type": "VcfAdapter",
         "vcfLocation": {
           "uri": "variant_effect_demo_data.vcf",
+          "locationType": "UriLocation"
+        }
+      },
+      "category": ["Variation"],
+      "assemblyNames": ["volvox"]
+    },
+    {
+      "type": "VariantTrack",
+      "trackId": "variant_effect_demo_jannovar",
+      "name": "Jannovar example",
+      "adapter": {
+        "type": "VcfAdapter",
+        "vcfLocation": {
+          "uri": "volvox_vcf_jannovar.vcf",
           "locationType": "UriLocation"
         }
       },
